Modified slightly so the publically available Tag is still sealed.
